DTFS2-4525 - fix/improve TFM/central API deals query with timestamp values

+++ trade-finance-manager-api/src/graphql/schemas/index.js
@@ -117,7 +117,7 @@ type DealEligibility {
   agentAddressTown: String
   agentName: String
   agentAlias: String
-  lastUpdated: Int
+  lastUpdated: Float
 }
 
 type FacilityProduct {
